setup features and screens folders
