+++ src/features/restaurants/components/restaurant-info.component.js
@@ -1,0 +1,4 @@
+import React from "react";
+import { Text } from "react-native";
+
+export const RestaurantInfo = () => <Text>Restaurant Info</Text>;
