+++ src/features/restaurants/screens/restaurants.screen.js
@@ -1,0 +1,31 @@
+import React from "react";
+import { Searchbar } from "react-native-paper";
+import { StatusBar, StyleSheet, SafeAreaView, View } from "react-native";
+
+import { RestaurantInfo } from "../components/restaurant-info.component";
+
+export const RestaurantsScreen = () => (
+  <SafeAreaView style={styles.container}>
+    <View style={styles.search}>
+      <Searchbar placeholder="Search for Places" />
+    </View>
+    <View style={styles.list}>
+      <RestaurantInfo />
+    </View>
+  </SafeAreaView>
+);
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    marginTop: StatusBar.currentHeight,
+  },
+  search: {
+    padding: 16,
+  },
+  list: {
+    flex: 1,
+    padding: 16,
+    backgroundColor: "blue",
+  },
+});
